ToDotNot unit test conversion - part 5.
