+++ src/Json.Schema.ToDotNet.UnitTests/DataModelGeneratorTests.cs
@@ -263,36 +263,13 @@ namespace Microsoft.Json.Schema.ToDotNet.UnitTests
         [Fact(DisplayName = "DataModelGenerator generates copyright notice")]
         public void GeneratesCopyrightAtTopOfFile()
         {
-            _settings.CopyrightNotice =
-@"// Copyright (c) 2016. All rights reserved.
-// Licensed under Apache 2.0 license.
-
-";
+            _settings.CopyrightNotice = TestUtil.ReadTestInputFile(ClassName, "CopyrightNotice.txt");
             var generator = new DataModelGenerator(_settings, _testFileSystem.FileSystem);
 
             JsonSchema schema = TestUtil.CreateSchemaFromTestDataFile("PropertyDescription");
 
-            const string Expected =
-@"// Copyright (c) 2016. All rights reserved.
-// Licensed under Apache 2.0 license.
+            string Expected = TestUtil.ReadTestInputFile(ClassName, "Expected.cs");
 
-using System;
-using System.CodeDom.Compiler;
-using System.Runtime.Serialization;
-
-namespace N
-{
-    [DataContract]
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    public partial class C
-    {
-        /// <summary>
-        /// An example property.
-        /// </summary>
-        [DataMember(Name = ""exampleProp"", IsRequired = false, EmitDefaultValue = false)]
-        public string ExampleProp { get; set; }
-    }
-}";
             string actual = generator.Generate(schema);
             actual.Should().Be(Expected);
         }
@@ -300,835 +277,16 @@ namespace N
         [Fact(DisplayName = "DataModelGenerator generates cloning code")]
         public void GeneratesCloningCode()
         {
-            JsonSchema schema = SchemaReader.ReadSchema(
-@"{
-  ""type"": ""object"",
-  ""properties"": {
-    ""integerProperty"": {
-      ""type"": ""integer"",
-      ""description"": ""An integer property.""
-    },
-    ""integerPropertyWithDefault"": {
-      ""type"": ""integer"",
-      ""description"": ""An integer property with a default value."",
-      ""default"": 42
-    },
-    ""numberProperty"": {
-      ""type"": ""number"",
-      ""description"": ""A number property.""
-    },
-    ""numberPropertyWithDefault"": {
-      ""type"": ""number"",
-      ""description"": ""A number property with a default value."",
-      ""default"": 42.1
-    },
-    ""stringProperty"": {
-      ""type"": ""string"",
-      ""description"": ""A string property.""
-    },
-    ""stringPropertyWithDefault"": {
-      ""type"": ""string"",
-      ""description"": ""A string property with a default value."",
-      ""default"": ""Don't panic.""
-    },
-    ""booleanProperty"": {
-      ""type"": ""boolean"",
-      ""description"": ""A Boolean property.""
-    },
-    ""booleanPropertyWithTrueDefault"": {
-      ""type"": ""boolean"",
-      ""description"": ""A Boolean property with a true default value."",
-      ""default"": true
-    },
-    ""booleanPropertyWithFalseDefault"": {
-      ""type"": ""boolean"",
-      ""description"": ""A Boolean property with a false default value."",
-      ""default"": false
-    },
-    ""enumeratedPropertyWithDefault"": {
-      ""description"": ""An enumerated property with a default value."",
-      ""enum"": [ ""red"", ""green"", ""blue"", ""black"", ""white"" ],
-      ""default"": ""green""
-    },
-    ""arrayProp"": {
-      ""type"": ""array"",
-      ""description"": ""An array property."",
-      ""items"": {
-        ""type"": ""number""
-      }
-    },
-    ""uriProp"": {
-      ""type"": ""string"",
-      ""description"": ""A Uri property."",
-      ""format"": ""uri""
-    },
-    ""dateTimeProp"": {
-      ""type"": ""string"",
-      ""description"": ""A DateTime property."",
-      ""format"": ""date-time""
-    },
-    ""referencedTypeProp"": {
-      ""$ref"": ""#/definitions/d""
-    },
-    ""arrayOfRefProp"": {
-      ""type"": ""array"",
-      ""description"": ""An array of a cloneable type."",
-      ""items"": {
-        ""$ref"": ""#/definitions/d""
-      }
-    },
-    ""arrayOfArrayProp"": {
-      ""type"": ""array"",
-      ""description"": ""An array of arrays."",
-      ""items"": {
-        ""type"": ""array"",
-        ""items"": {
-          ""$ref"": ""#/definitions/d""
-        }
-      }
-    },
-    ""dictionaryProp"": {
-      ""description"": ""A dictionary property."",
-      ""type"": ""object""
-    },
-    ""dictionaryWithPrimitiveSchemaProp"": {
-      ""description"": ""A dictionary property whose values are defined by a primitive additionalProperties schema."",
-      ""type"": ""object"",
-      ""additionalProperties"": {
-        ""type"": ""number""
-      }
-    },
-    ""dictionaryWithObjectSchemaProp"": {
-      ""description"": ""A dictionary property whose values are defined by an object-valued additionalProperties schema."",
-      ""type"": ""object"",
-      ""additionalProperties"": {
-        ""$ref"": ""#/definitions/d""
-      }
-    },
-    ""dictionaryWithObjectArraySchemaProp"": {
-      ""description"": ""A dictionary property whose values are defined by an array-of-object-valued additionalProperties schema."",
-      ""type"": ""object"",
-      ""additionalProperties"": {
-        ""type"": ""array"",
-        ""items"": {
-          ""$ref"": ""#/definitions/d""
-        }
-      }
-    },
-    ""dictionaryWithUriKeyProp"": {
-      ""description"": ""A dictionary property whose keys are Uris."",
-      ""type"": ""object"",
-      ""additionalProperties"": {
-        ""$ref"": ""#/definitions/d""
-      }
-    },
-    ""dictionaryWithHintedValueProp"": {
-      ""description"": ""A dictionary property whose value type is hinted."",
-      ""type"": ""object""
-    }
-  },
-  ""definitions"": {
-    ""d"": {
-      ""type"": ""object""
-    }
-  }
-}", TestUtil.TestFilePath);
+            string SchemaText = TestUtil.ReadTestInputFile(ClassName, "Schema.json");
+            JsonSchema schema = SchemaReader.ReadSchema(SchemaText, TestUtil.TestFilePath);
+
+            string HintsText = TestUtil.ReadTestInputFile(ClassName, "CodeGenHints.json");
+            string ExpectedClass = TestUtil.ReadTestInputFile(ClassName, "ExpectedClass.cs");
+            string ExpectedSyntaxInterface = TestUtil.ReadTestInputFile(ClassName, "ExpectedSyntaxInterface.cs");
+            string ExpectedKindEnum = TestUtil.ReadTestInputFile(ClassName, "ExpectedKindEnum.cs");
+            string ExpectedRewritingVisitor = TestUtil.ReadTestInputFile(ClassName, "ExpectedRewritingVisitor.cs");
+            string ExpectedEnumType = TestUtil.ReadTestInputFile(ClassName, "ExpectedEnumType.cs");
 
-            const string HintsText =
-@"{
-  ""C.EnumeratedPropertyWithDefault"": [
-    {
-      ""kind"": ""EnumHint"",
-      ""arguments"": {
-        ""typeName"": ""Color"",
-        ""description"": ""Some colors.""
-      }
-    }
-  ],
-  ""C.DictionaryProp"": [
-    {
-      ""kind"": ""DictionaryHint""
-    }
-  ],
-  ""C.DictionaryWithPrimitiveSchemaProp"": [
-    {
-      ""kind"": ""DictionaryHint""
-    }
-  ],
-  ""C.DictionaryWithObjectSchemaProp"": [
-    {
-      ""kind"": ""DictionaryHint""
-    }
-  ],
-  ""C.DictionaryWithObjectArraySchemaProp"": [
-    {
-      ""kind"": ""DictionaryHint""
-    }
-  ],
-  ""C.DictionaryWithUriKeyProp"": [
-    {
-      ""kind"": ""DictionaryHint"",
-      ""arguments"": {
-        ""keyTypeName"": ""Uri""
-      }
-    }
-  ],
-  ""C.DictionaryWithHintedValueProp"": [
-    {
-      ""kind"": ""DictionaryHint"",
-      ""arguments"": {
-        ""valueTypeName"": ""V"",
-        ""comparisonKind"": ""ObjectEquals"",
-        ""hashKind"": ""ScalarReferenceType"",
-        ""initializationKind"": ""SimpleAssign""
-      }
-    }
-  ]
-}";
-
-            const string ExpectedClass =
-@"using System;
-using System.CodeDom.Compiler;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Runtime.Serialization;
-using Newtonsoft.Json;
-
-namespace N
-{
-    [DataContract]
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    public partial class C : ISNode
-    {
-        /// <summary>
-        /// Gets a value indicating the type of object implementing <see cref=""ISNode"" />.
-        /// </summary>
-        public SNodeKind SNodeKind
-        {
-            get
-            {
-                return SNodeKind.C;
-            }
-        }
-
-        /// <summary>
-        /// An integer property.
-        /// </summary>
-        [DataMember(Name = ""integerProperty"", IsRequired = false, EmitDefaultValue = false)]
-        public int IntegerProperty { get; set; }
-
-        /// <summary>
-        /// An integer property with a default value.
-        /// </summary>
-        [DataMember(Name = ""integerPropertyWithDefault"", IsRequired = false, EmitDefaultValue = false)]
-        [DefaultValue(42)]
-        [JsonProperty(DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
-        public int IntegerPropertyWithDefault { get; set; }
-
-        /// <summary>
-        /// A number property.
-        /// </summary>
-        [DataMember(Name = ""numberProperty"", IsRequired = false, EmitDefaultValue = false)]
-        public double NumberProperty { get; set; }
-
-        /// <summary>
-        /// A number property with a default value.
-        /// </summary>
-        [DataMember(Name = ""numberPropertyWithDefault"", IsRequired = false, EmitDefaultValue = false)]
-        [DefaultValue(42.1)]
-        [JsonProperty(DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
-        public double NumberPropertyWithDefault { get; set; }
-
-        /// <summary>
-        /// A string property.
-        /// </summary>
-        [DataMember(Name = ""stringProperty"", IsRequired = false, EmitDefaultValue = false)]
-        public string StringProperty { get; set; }
-
-        /// <summary>
-        /// A string property with a default value.
-        /// </summary>
-        [DataMember(Name = ""stringPropertyWithDefault"", IsRequired = false, EmitDefaultValue = false)]
-        [DefaultValue(""Don't panic."")]
-        [JsonProperty(DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
-        public string StringPropertyWithDefault { get; set; }
-
-        /// <summary>
-        /// A Boolean property.
-        /// </summary>
-        [DataMember(Name = ""booleanProperty"", IsRequired = false, EmitDefaultValue = false)]
-        public bool BooleanProperty { get; set; }
-
-        /// <summary>
-        /// A Boolean property with a true default value.
-        /// </summary>
-        [DataMember(Name = ""booleanPropertyWithTrueDefault"", IsRequired = false, EmitDefaultValue = false)]
-        [DefaultValue(true)]
-        [JsonProperty(DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
-        public bool BooleanPropertyWithTrueDefault { get; set; }
-
-        /// <summary>
-        /// A Boolean property with a false default value.
-        /// </summary>
-        [DataMember(Name = ""booleanPropertyWithFalseDefault"", IsRequired = false, EmitDefaultValue = false)]
-        [DefaultValue(false)]
-        [JsonProperty(DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
-        public bool BooleanPropertyWithFalseDefault { get; set; }
-
-        /// <summary>
-        /// An enumerated property with a default value.
-        /// </summary>
-        [DataMember(Name = ""enumeratedPropertyWithDefault"", IsRequired = false, EmitDefaultValue = false)]
-        [DefaultValue(Color.Green)]
-        [JsonProperty(DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
-        public Color EnumeratedPropertyWithDefault { get; set; }
-
-        /// <summary>
-        /// An array property.
-        /// </summary>
-        [DataMember(Name = ""arrayProp"", IsRequired = false, EmitDefaultValue = false)]
-        public IList<double> ArrayProp { get; set; }
-
-        /// <summary>
-        /// A Uri property.
-        /// </summary>
-        [DataMember(Name = ""uriProp"", IsRequired = false, EmitDefaultValue = false)]
-        public Uri UriProp { get; set; }
-
-        /// <summary>
-        /// A DateTime property.
-        /// </summary>
-        [DataMember(Name = ""dateTimeProp"", IsRequired = false, EmitDefaultValue = false)]
-        public DateTime DateTimeProp { get; set; }
-        [DataMember(Name = ""referencedTypeProp"", IsRequired = false, EmitDefaultValue = false)]
-        public D ReferencedTypeProp { get; set; }
-
-        /// <summary>
-        /// An array of a cloneable type.
-        /// </summary>
-        [DataMember(Name = ""arrayOfRefProp"", IsRequired = false, EmitDefaultValue = false)]
-        public IList<D> ArrayOfRefProp { get; set; }
-
-        /// <summary>
-        /// An array of arrays.
-        /// </summary>
-        [DataMember(Name = ""arrayOfArrayProp"", IsRequired = false, EmitDefaultValue = false)]
-        public IList<IList<D>> ArrayOfArrayProp { get; set; }
-
-        /// <summary>
-        /// A dictionary property.
-        /// </summary>
-        [DataMember(Name = ""dictionaryProp"", IsRequired = false, EmitDefaultValue = false)]
-        public IDictionary<string, string> DictionaryProp { get; set; }
-
-        /// <summary>
-        /// A dictionary property whose values are defined by a primitive additionalProperties schema.
-        /// </summary>
-        [DataMember(Name = ""dictionaryWithPrimitiveSchemaProp"", IsRequired = false, EmitDefaultValue = false)]
-        public IDictionary<string, double> DictionaryWithPrimitiveSchemaProp { get; set; }
-
-        /// <summary>
-        /// A dictionary property whose values are defined by an object-valued additionalProperties schema.
-        /// </summary>
-        [DataMember(Name = ""dictionaryWithObjectSchemaProp"", IsRequired = false, EmitDefaultValue = false)]
-        public IDictionary<string, D> DictionaryWithObjectSchemaProp { get; set; }
-
-        /// <summary>
-        /// A dictionary property whose values are defined by an array-of-object-valued additionalProperties schema.
-        /// </summary>
-        [DataMember(Name = ""dictionaryWithObjectArraySchemaProp"", IsRequired = false, EmitDefaultValue = false)]
-        public IDictionary<string, IList<D>> DictionaryWithObjectArraySchemaProp { get; set; }
-
-        /// <summary>
-        /// A dictionary property whose keys are Uris.
-        /// </summary>
-        [DataMember(Name = ""dictionaryWithUriKeyProp"", IsRequired = false, EmitDefaultValue = false)]
-        public IDictionary<Uri, D> DictionaryWithUriKeyProp { get; set; }
-
-        /// <summary>
-        /// A dictionary property whose value type is hinted.
-        /// </summary>
-        [DataMember(Name = ""dictionaryWithHintedValueProp"", IsRequired = false, EmitDefaultValue = false)]
-        public IDictionary<string, V> DictionaryWithHintedValueProp { get; set; }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref=""C"" /> class.
-        /// </summary>
-        public C()
-        {
-            IntegerPropertyWithDefault = 42;
-            NumberPropertyWithDefault = 42.1;
-            StringPropertyWithDefault = ""Don't panic."";
-            BooleanPropertyWithTrueDefault = true;
-            BooleanPropertyWithFalseDefault = false;
-            EnumeratedPropertyWithDefault = Color.Green;
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref=""C"" /> class from the supplied values.
-        /// </summary>
-        /// <param name=""integerProperty"">
-        /// An initialization value for the <see cref=""P:IntegerProperty"" /> property.
-        /// </param>
-        /// <param name=""integerPropertyWithDefault"">
-        /// An initialization value for the <see cref=""P:IntegerPropertyWithDefault"" /> property.
-        /// </param>
-        /// <param name=""numberProperty"">
-        /// An initialization value for the <see cref=""P:NumberProperty"" /> property.
-        /// </param>
-        /// <param name=""numberPropertyWithDefault"">
-        /// An initialization value for the <see cref=""P:NumberPropertyWithDefault"" /> property.
-        /// </param>
-        /// <param name=""stringProperty"">
-        /// An initialization value for the <see cref=""P:StringProperty"" /> property.
-        /// </param>
-        /// <param name=""stringPropertyWithDefault"">
-        /// An initialization value for the <see cref=""P:StringPropertyWithDefault"" /> property.
-        /// </param>
-        /// <param name=""booleanProperty"">
-        /// An initialization value for the <see cref=""P:BooleanProperty"" /> property.
-        /// </param>
-        /// <param name=""booleanPropertyWithTrueDefault"">
-        /// An initialization value for the <see cref=""P:BooleanPropertyWithTrueDefault"" /> property.
-        /// </param>
-        /// <param name=""booleanPropertyWithFalseDefault"">
-        /// An initialization value for the <see cref=""P:BooleanPropertyWithFalseDefault"" /> property.
-        /// </param>
-        /// <param name=""enumeratedPropertyWithDefault"">
-        /// An initialization value for the <see cref=""P:EnumeratedPropertyWithDefault"" /> property.
-        /// </param>
-        /// <param name=""arrayProp"">
-        /// An initialization value for the <see cref=""P:ArrayProp"" /> property.
-        /// </param>
-        /// <param name=""uriProp"">
-        /// An initialization value for the <see cref=""P:UriProp"" /> property.
-        /// </param>
-        /// <param name=""dateTimeProp"">
-        /// An initialization value for the <see cref=""P:DateTimeProp"" /> property.
-        /// </param>
-        /// <param name=""referencedTypeProp"">
-        /// An initialization value for the <see cref=""P:ReferencedTypeProp"" /> property.
-        /// </param>
-        /// <param name=""arrayOfRefProp"">
-        /// An initialization value for the <see cref=""P:ArrayOfRefProp"" /> property.
-        /// </param>
-        /// <param name=""arrayOfArrayProp"">
-        /// An initialization value for the <see cref=""P:ArrayOfArrayProp"" /> property.
-        /// </param>
-        /// <param name=""dictionaryProp"">
-        /// An initialization value for the <see cref=""P:DictionaryProp"" /> property.
-        /// </param>
-        /// <param name=""dictionaryWithPrimitiveSchemaProp"">
-        /// An initialization value for the <see cref=""P:DictionaryWithPrimitiveSchemaProp"" /> property.
-        /// </param>
-        /// <param name=""dictionaryWithObjectSchemaProp"">
-        /// An initialization value for the <see cref=""P:DictionaryWithObjectSchemaProp"" /> property.
-        /// </param>
-        /// <param name=""dictionaryWithObjectArraySchemaProp"">
-        /// An initialization value for the <see cref=""P:DictionaryWithObjectArraySchemaProp"" /> property.
-        /// </param>
-        /// <param name=""dictionaryWithUriKeyProp"">
-        /// An initialization value for the <see cref=""P:DictionaryWithUriKeyProp"" /> property.
-        /// </param>
-        /// <param name=""dictionaryWithHintedValueProp"">
-        /// An initialization value for the <see cref=""P:DictionaryWithHintedValueProp"" /> property.
-        /// </param>
-        public C(int integerProperty, int integerPropertyWithDefault, double numberProperty, double numberPropertyWithDefault, string stringProperty, string stringPropertyWithDefault, bool booleanProperty, bool booleanPropertyWithTrueDefault, bool booleanPropertyWithFalseDefault, Color enumeratedPropertyWithDefault, IEnumerable<double> arrayProp, Uri uriProp, DateTime dateTimeProp, D referencedTypeProp, IEnumerable<D> arrayOfRefProp, IEnumerable<IEnumerable<D>> arrayOfArrayProp, IDictionary<string, string> dictionaryProp, IDictionary<string, double> dictionaryWithPrimitiveSchemaProp, IDictionary<string, D> dictionaryWithObjectSchemaProp, IDictionary<string, IList<D>> dictionaryWithObjectArraySchemaProp, IDictionary<Uri, D> dictionaryWithUriKeyProp, IDictionary<string, V> dictionaryWithHintedValueProp)
-        {
-            Init(integerProperty, integerPropertyWithDefault, numberProperty, numberPropertyWithDefault, stringProperty, stringPropertyWithDefault, booleanProperty, booleanPropertyWithTrueDefault, booleanPropertyWithFalseDefault, enumeratedPropertyWithDefault, arrayProp, uriProp, dateTimeProp, referencedTypeProp, arrayOfRefProp, arrayOfArrayProp, dictionaryProp, dictionaryWithPrimitiveSchemaProp, dictionaryWithObjectSchemaProp, dictionaryWithObjectArraySchemaProp, dictionaryWithUriKeyProp, dictionaryWithHintedValueProp);
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref=""C"" /> class from the specified instance.
-        /// </summary>
-        /// <param name=""other"">
-        /// The instance from which the new instance is to be initialized.
-        /// </param>
-        /// <exception cref=""ArgumentNullException"">
-        /// Thrown if <paramref name=""other"" /> is null.
-        /// </exception>
-        public C(C other)
-        {
-            if (other == null)
-            {
-                throw new ArgumentNullException(nameof(other));
-            }
-
-            Init(other.IntegerProperty, other.IntegerPropertyWithDefault, other.NumberProperty, other.NumberPropertyWithDefault, other.StringProperty, other.StringPropertyWithDefault, other.BooleanProperty, other.BooleanPropertyWithTrueDefault, other.BooleanPropertyWithFalseDefault, other.EnumeratedPropertyWithDefault, other.ArrayProp, other.UriProp, other.DateTimeProp, other.ReferencedTypeProp, other.ArrayOfRefProp, other.ArrayOfArrayProp, other.DictionaryProp, other.DictionaryWithPrimitiveSchemaProp, other.DictionaryWithObjectSchemaProp, other.DictionaryWithObjectArraySchemaProp, other.DictionaryWithUriKeyProp, other.DictionaryWithHintedValueProp);
-        }
-
-        ISNode ISNode.DeepClone()
-        {
-            return DeepCloneCore();
-        }
-
-        /// <summary>
-        /// Creates a deep copy of this instance.
-        /// </summary>
-        public C DeepClone()
-        {
-            return (C)DeepCloneCore();
-        }
-
-        private ISNode DeepCloneCore()
-        {
-            return new C(this);
-        }
-
-        private void Init(int integerProperty, int integerPropertyWithDefault, double numberProperty, double numberPropertyWithDefault, string stringProperty, string stringPropertyWithDefault, bool booleanProperty, bool booleanPropertyWithTrueDefault, bool booleanPropertyWithFalseDefault, Color enumeratedPropertyWithDefault, IEnumerable<double> arrayProp, Uri uriProp, DateTime dateTimeProp, D referencedTypeProp, IEnumerable<D> arrayOfRefProp, IEnumerable<IEnumerable<D>> arrayOfArrayProp, IDictionary<string, string> dictionaryProp, IDictionary<string, double> dictionaryWithPrimitiveSchemaProp, IDictionary<string, D> dictionaryWithObjectSchemaProp, IDictionary<string, IList<D>> dictionaryWithObjectArraySchemaProp, IDictionary<Uri, D> dictionaryWithUriKeyProp, IDictionary<string, V> dictionaryWithHintedValueProp)
-        {
-            IntegerProperty = integerProperty;
-            IntegerPropertyWithDefault = integerPropertyWithDefault;
-            NumberProperty = numberProperty;
-            NumberPropertyWithDefault = numberPropertyWithDefault;
-            StringProperty = stringProperty;
-            StringPropertyWithDefault = stringPropertyWithDefault;
-            BooleanProperty = booleanProperty;
-            BooleanPropertyWithTrueDefault = booleanPropertyWithTrueDefault;
-            BooleanPropertyWithFalseDefault = booleanPropertyWithFalseDefault;
-            EnumeratedPropertyWithDefault = enumeratedPropertyWithDefault;
-            if (arrayProp != null)
-            {
-                var destination_0 = new List<double>();
-                foreach (var value_0 in arrayProp)
-                {
-                    destination_0.Add(value_0);
-                }
-
-                ArrayProp = destination_0;
-            }
-
-            if (uriProp != null)
-            {
-                UriProp = new Uri(uriProp.OriginalString, uriProp.IsAbsoluteUri ? UriKind.Absolute : UriKind.Relative);
-            }
-
-            DateTimeProp = dateTimeProp;
-            if (referencedTypeProp != null)
-            {
-                ReferencedTypeProp = new D(referencedTypeProp);
-            }
-
-            if (arrayOfRefProp != null)
-            {
-                var destination_1 = new List<D>();
-                foreach (var value_1 in arrayOfRefProp)
-                {
-                    if (value_1 == null)
-                    {
-                        destination_1.Add(null);
-                    }
-                    else
-                    {
-                        destination_1.Add(new D(value_1));
-                    }
-                }
-
-                ArrayOfRefProp = destination_1;
-            }
-
-            if (arrayOfArrayProp != null)
-            {
-                var destination_2 = new List<IList<D>>();
-                foreach (var value_2 in arrayOfArrayProp)
-                {
-                    if (value_2 == null)
-                    {
-                        destination_2.Add(null);
-                    }
-                    else
-                    {
-                        var destination_3 = new List<D>();
-                        foreach (var value_3 in value_2)
-                        {
-                            if (value_3 == null)
-                            {
-                                destination_3.Add(null);
-                            }
-                            else
-                            {
-                                destination_3.Add(new D(value_3));
-                            }
-                        }
-
-                        destination_2.Add(destination_3);
-                    }
-                }
-
-                ArrayOfArrayProp = destination_2;
-            }
-
-            if (dictionaryProp != null)
-            {
-                DictionaryProp = new Dictionary<string, string>(dictionaryProp);
-            }
-
-            if (dictionaryWithPrimitiveSchemaProp != null)
-            {
-                DictionaryWithPrimitiveSchemaProp = new Dictionary<string, double>(dictionaryWithPrimitiveSchemaProp);
-            }
-
-            if (dictionaryWithObjectSchemaProp != null)
-            {
-                DictionaryWithObjectSchemaProp = new Dictionary<string, D>();
-                foreach (var value_4 in dictionaryWithObjectSchemaProp)
-                {
-                    DictionaryWithObjectSchemaProp.Add(value_4.Key, new D(value_4.Value));
-                }
-            }
-
-            if (dictionaryWithObjectArraySchemaProp != null)
-            {
-                DictionaryWithObjectArraySchemaProp = new Dictionary<string, IList<D>>();
-                foreach (var value_5 in dictionaryWithObjectArraySchemaProp)
-                {
-                    var destination_4 = new List<D>();
-                    foreach (var value_6 in value_5.Value)
-                    {
-                        if (value_6 == null)
-                        {
-                            destination_4.Add(null);
-                        }
-                        else
-                        {
-                            destination_4.Add(new D(value_6));
-                        }
-                    }
-
-                    DictionaryWithObjectArraySchemaProp.Add(value_5.Key, destination_4);
-                }
-            }
-
-            if (dictionaryWithUriKeyProp != null)
-            {
-                DictionaryWithUriKeyProp = new Dictionary<Uri, D>();
-                foreach (var value_7 in dictionaryWithUriKeyProp)
-                {
-                    DictionaryWithUriKeyProp.Add(value_7.Key, new D(value_7.Value));
-                }
-            }
-
-            if (dictionaryWithHintedValueProp != null)
-            {
-                DictionaryWithHintedValueProp = new Dictionary<string, V>(dictionaryWithHintedValueProp);
-            }
-        }
-    }
-}";
-            const string ExpectedSyntaxInterface =
-@"using System.CodeDom.Compiler;
-
-namespace N
-{
-    /// <summary>
-    /// An interface for all types generated from the S schema.
-    /// </summary>
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    public interface ISNode
-    {
-        /// <summary>
-        /// Gets a value indicating the type of object implementing <see cref=""ISNode"" />.
-        /// </summary>
-        SNodeKind SNodeKind { get; }
-
-        /// <summary>
-        /// Makes a deep copy of this instance.
-        /// </summary>
-        ISNode DeepClone();
-    }
-}";
-            const string ExpectedKindEnum =
-@"using System.CodeDom.Compiler;
-
-namespace N
-{
-    /// <summary>
-    /// A set of values for all the types that implement <see cref=""ISNode"" />.
-    /// </summary>
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    public enum SNodeKind
-    {
-        /// <summary>
-        /// An uninitialized kind.
-        /// </summary>
-        None,
-        /// <summary>
-        /// A value indicating that the <see cref=""ISNode"" /> object is of type <see cref=""C"" />.
-        /// </summary>
-        C,
-        /// <summary>
-        /// A value indicating that the <see cref=""ISNode"" /> object is of type <see cref=""D"" />.
-        /// </summary>
-        D
-    }
-}";
-            const string ExpectedRewritingVisitor =
-@"using System;
-using System.CodeDom.Compiler;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace N
-{
-    /// <summary>
-    /// Rewriting visitor for the S object model.
-    /// </summary>
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    public abstract class SRewritingVisitor
-    {
-        /// <summary>
-        /// Starts a rewriting visit of a node in the S object model.
-        /// </summary>
-        /// <param name=""node"">
-        /// The node to rewrite.
-        /// </param>
-        /// <returns>
-        /// A rewritten instance of the node.
-        /// </returns>
-        public virtual object Visit(ISNode node)
-        {
-            return this.VisitActual(node);
-        }
-
-        /// <summary>
-        /// Visits and rewrites a node in the S object model.
-        /// </summary>
-        /// <param name=""node"">
-        /// The node to rewrite.
-        /// </param>
-        /// <returns>
-        /// A rewritten instance of the node.
-        /// </returns>
-        public virtual object VisitActual(ISNode node)
-        {
-            if (node == null)
-            {
-                throw new ArgumentNullException(""node"");
-            }
-
-            switch (node.SNodeKind)
-            {
-                case SNodeKind.C:
-                    return VisitC((C)node);
-                case SNodeKind.D:
-                    return VisitD((D)node);
-                default:
-                    return node;
-            }
-        }
-
-        private T VisitNullChecked<T>(T node) where T : class, ISNode
-        {
-            if (node == null)
-            {
-                return null;
-            }
-
-            return (T)Visit(node);
-        }
-
-        public virtual C VisitC(C node)
-        {
-            if (node != null)
-            {
-                node.ReferencedTypeProp = VisitNullChecked(node.ReferencedTypeProp);
-                if (node.ArrayOfRefProp != null)
-                {
-                    for (int index_0 = 0; index_0 < node.ArrayOfRefProp.Count; ++index_0)
-                    {
-                        node.ArrayOfRefProp[index_0] = VisitNullChecked(node.ArrayOfRefProp[index_0]);
-                    }
-                }
-
-                if (node.ArrayOfArrayProp != null)
-                {
-                    for (int index_0 = 0; index_0 < node.ArrayOfArrayProp.Count; ++index_0)
-                    {
-                        var value_0 = node.ArrayOfArrayProp[index_0];
-                        if (value_0 != null)
-                        {
-                            for (int index_1 = 0; index_1 < value_0.Count; ++index_1)
-                            {
-                                value_0[index_1] = VisitNullChecked(value_0[index_1]);
-                            }
-                        }
-                    }
-                }
-
-                if (node.DictionaryWithObjectSchemaProp != null)
-                {
-                    var keys = node.DictionaryWithObjectSchemaProp.Keys.ToArray();
-                    foreach (var key in keys)
-                    {
-                        var value = node.DictionaryWithObjectSchemaProp[key];
-                        if (value != null)
-                        {
-                            node.DictionaryWithObjectSchemaProp[key] = VisitNullChecked(value);
-                        }
-                    }
-                }
-
-                if (node.DictionaryWithObjectArraySchemaProp != null)
-                {
-                    var keys = node.DictionaryWithObjectArraySchemaProp.Keys.ToArray();
-                    foreach (var key in keys)
-                    {
-                        var value = node.DictionaryWithObjectArraySchemaProp[key];
-                        if (value != null)
-                        {
-                            for (int index_0 = 0; index_0 < node.DictionaryWithObjectArraySchemaProp[key].Count; ++index_0)
-                            {
-                                node.DictionaryWithObjectArraySchemaProp[key][index_0] = VisitNullChecked(node.DictionaryWithObjectArraySchemaProp[key][index_0]);
-                            }
-                        }
-                    }
-                }
-
-                if (node.DictionaryWithUriKeyProp != null)
-                {
-                    var keys = node.DictionaryWithUriKeyProp.Keys.ToArray();
-                    foreach (var key in keys)
-                    {
-                        var value = node.DictionaryWithUriKeyProp[key];
-                        if (value != null)
-                        {
-                            node.DictionaryWithUriKeyProp[key] = VisitNullChecked(value);
-                        }
-                    }
-                }
-            }
-
-            return node;
-        }
-
-        public virtual D VisitD(D node)
-        {
-            if (node != null)
-            {
-            }
-
-            return node;
-        }
-    }
-}";
-            const string ExpectedEnumType =
-@"using System.CodeDom.Compiler;
-
-namespace N
-{
-    /// <summary>
-    /// Some colors.
-    /// </summary>
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """+ VersionConstants.FileVersion + @""")]
-    public enum Color
-    {
-        Red,
-        Green,
-        Blue,
-        Black,
-        White
-    }
-}";
             _settings.GenerateCloningCode = true;
             _settings.HintDictionary = new HintDictionary(HintsText);
             var generator = new DataModelGenerator(_settings, _testFileSystem.FileSystem);
