+++ src/Json.Schema.ToDotNet.UnitTests/TestData/DataModelGeneratorTests/GeneratesCloningCode/ExpectedEnumType.cs
@@ -1,0 +1,17 @@
+using System.CodeDom.Compiler;
+
+namespace N
+{
+    /// <summary>
+    /// Some colors.
+    /// </summary>
+    [GeneratedCode("Microsoft.Json.Schema.ToDotNet", "$JSchemaFileVersion$")]
+    public enum Color
+    {
+        Red,
+        Green,
+        Blue,
+        Black,
+        White
+    }
+}
