+++ src/Json.Schema.ToDotNet.UnitTests/TestData/DataModelGeneratorTests/GeneratesCopyrightAtTopOfFile/CopyrightNotice.txt
@@ -1,0 +1,3 @@
+// Copyright (c) 2016. All rights reserved.
+// Licensed under Apache 2.0 license.
+
